+++ src/cpp/core/system/PosixSystem.cpp
@@ -2472,9 +2472,10 @@ Error permanentlyDropPriv(const std::string& newUsername)
    if (error)
       return error;
 
-   // before changing the process user, ensure it becomes the new
-   // owner of any file logs so we can ensure that we can keep writing to them
-   core::log::refreshAllLogDestinations(core::log::RefreshParams{ user });
+   // Refresh the destinations without providing a user since we are have the parent processes's file destinations 
+   // and don't want to change the ownership of the log files to newUsername. We do still want to make sure we are set up to log in case
+   // there are errors before we call ::execve - in particular, for syslog
+   core::log::refreshAllLogDestinations();
 
    // clear error state
    errno = 0;