+++ src/node/desktop/src/main/secondary-window.ts
@@ -14,7 +14,7 @@
  *
  */
 
-import { WebContents } from 'electron';
+import { BrowserWindow, WebContents } from 'electron';
 import { DesktopBrowserWindow } from './desktop-browser-window';
 
 export class SecondaryWindow extends DesktopBrowserWindow {
@@ -24,8 +24,10 @@ export class SecondaryWindow extends DesktopBrowserWindow {
     baseUrl?: string,
     parent?: DesktopBrowserWindow,
     opener?: WebContents,
-    allowExternalNavigate = false
+    allowExternalNavigate = false,
+    existingWindow?: BrowserWindow
   ) {
-    super(showToolbar, true, name, baseUrl, parent, opener, allowExternalNavigate);
+    super(showToolbar, true, name, baseUrl, parent, opener,
+      allowExternalNavigate, undefined, existingWindow);
   }
 }