+++ src/node/desktop/src/main/utils.ts
@@ -17,13 +17,16 @@ import fs, { existsSync } from 'fs';
 import os from 'os';
 import path from 'path';
 import { sep } from 'path';
-import { app, FileFilter, WebContents } from 'electron';
+import { app, BrowserWindow, FileFilter, WebContents } from 'electron';
+import http from 'http';
 
 import { Xdg } from '../core/xdg';
 import { getenv, setenv } from '../core/environment';
 import { FilePath } from '../core/file-path';
 import { logger } from '../core/logger';
 import { userHomePath } from '../core/user';
+import { WaitResult, WaitTimeoutFn, waitWithTimeout } from '../core/wait-utils';
+import { Err } from '../core/err';
 
 import { productInfo } from './product-info';
 import { MainWindow } from './main-window';
@@ -312,4 +315,62 @@ export function filterFromQFileDialogFilter(qtFilters: string): FileFilter[] {
     result.push({ name: name, extensions: exts });
   }
   return result;
+}
+
+/**
+ * Wait for a URL to respond, with retries and timeout
+ */
+export async function waitForUrlWithTimeout(
+  url: string,
+  initialWaitMs: number,
+  incrementWaitMs: number,
+  maxWaitSec: number
+): Promise<Err> {
+
+  const checkReady: WaitTimeoutFn = async () => {
+    return new Promise((resolve) => {
+      http.get(url, (res) => {
+        res.resume(); // consume response data to free up memory
+        resolve(new WaitResult('WaitSuccess'));
+      }).on('error', (e) => {
+        logger().logDebug(`Connection to ${url} failed: ${e.message}`);
+        resolve(new WaitResult('WaitContinue'));
+      });
+    });
+  };
+
+  return waitWithTimeout(checkReady, initialWaitMs, incrementWaitMs, maxWaitSec);
+}
+
+export function raiseAndActivateWindow(window: BrowserWindow): void {
+  if (window.isMinimized()) {
+    window.restore();
+  }
+  window.focus();
+}
+
+export function getDpiZoomScaling(): number {
+  // TODO: because Qt is already high-DPI aware and automatically
+  // scales in most scenarios, we no longer need to detect and
+  // apply a custom scale -- but more testing is warranted
+  return 1.0;
+}
+
+/**
+ * Determine if given host is considered safe to load in an IDE window.
+ */
+export function isSafeHost(host: string): boolean {
+  const safeHosts = [
+    '.youtube.com',
+    '.vimeo.com',
+    '.c9.ms',
+    '.google.com'
+  ];
+
+  for (const safeHost of safeHosts) {
+    if (host.endsWith(safeHost)) {
+      return true;
+    }
+  }
+  return false;
 }