+++ NEWS.md
@@ -21,7 +21,8 @@
 
 ### RStudio Workbench
 
-* Load-balancing configuration is now automatic and doesn't require editing config files (Pro #2167)
+* Load-balancing configuration is now automatic, stored in the internal database, and doesn't require editing config files. Added supporting `rstudio-server list-nodes` command to show users the node's status according to the database. (Pro #2167)
+* Added `rstudio-server delete-node` command to allow users to remove a load-balancing node from the database (Pro #2800)
 * Added AWS Cognito support to OpenID integration (Pro #2313)
 * Add file uploads and downloads to session audit log (Pro #2226)
 * RStudio Workbench can now run without root access in a single-user mode (Pro #9496)
@@ -65,7 +66,6 @@
 * Support highlight of 'css' and 'asis' chunks in R Markdown documents (#4821)
 * Improved ordering of completion results within `library()` calls (#9293)
 * Syntax support for embedded knitr chunks (#9579)
-* Added support for publishing Quarto documents and websites (#9556)
 * Add option to synchronize the Files pane with the current working directory in R (#4615)
 * Add new *Set Working Directory* command to context menu for source files (#6781)
 * Local background jobs can now be replayed (#5548)